Move store utils to component library utilities
